Remove NetworkDatabase trait, collate netdb errors
